Added namespace to Notifications tests
