+++ tests/Assets/Notifications/CustomNotifiableStub.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace LaravelDoctrineTest\ORM\Assets\Notifications;
+
+use LaravelDoctrine\ORM\Notifications\Notifiable;
+
+class CustomNotifiableStub
+{
+    use Notifiable;
+
+    public function routeNotificationForDoctrine()
+    {
+        return 'custom';
+    }
+}
