+++ tests/Assets/Notifications/NotifiableStub.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace LaravelDoctrineTest\ORM\Assets\Notifications;
+
+use LaravelDoctrine\ORM\Notifications\Notifiable;
+
+class NotifiableStub
+{
+    use Notifiable;
+}
+
