+++ tests/Assets/Notifications/NotificationDatabaseStub.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace LaravelDoctrineTest\ORM\Assets\Notifications;
+
+class NotificationDatabaseStub extends \Illuminate\Notifications\Notification
+{
+    public function toDatabase()
+    {
+        return (new \LaravelDoctrine\ORM\Notifications\Notification());
+    }
+}
