+++ tests/Assets/Notifications/NotificationInvalidStub.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace LaravelDoctrineTest\ORM\Assets\Notifications;
+
+class NotificationInvalidStub extends \Illuminate\Notifications\Notification
+{
+}
