+++ tests/Assets/Notifications/NotificationStub.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace LaravelDoctrineTest\ORM\Assets\Notifications;
+
+class NotificationStub extends \Illuminate\Notifications\Notification
+{
+    public function toEntity()
+    {
+        return (new \LaravelDoctrine\ORM\Notifications\Notification());
+    }
+}
